Refactor to allow for two apis
