--- consonance-webservice/src/main/java/io/swagger/api/ApiException.java
@@ -1,10 +1,0 @@
-package io.swagger.api;
-
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2016-06-29T19:50:58.742Z")
-public class ApiException extends Exception{
-	private int code;
-	public ApiException (int code, String msg) {
-		super(msg);
-		this.code = code;
-	}
-}
