--- consonance-webservice/src/main/java/io/swagger/api/JacksonJsonProvider.java
@@ -1,19 +1,0 @@
-package io.swagger.api;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.jaxrs.json.JacksonJaxbJsonProvider;
-import io.swagger.util.Json;
-
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.ext.Provider;
-
-@Provider
-@Produces({MediaType.APPLICATION_JSON})
-public class JacksonJsonProvider extends JacksonJaxbJsonProvider {
-    private static ObjectMapper commonMapper = Json.mapper();
-
-    public JacksonJsonProvider() {
-        super.setMapper(commonMapper);
-    }
-}
