--- consonance-webservice/src/main/java/io/swagger/api/NotFoundException.java
@@ -1,10 +1,0 @@
-package io.swagger.api;
-
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2016-06-29T19:50:58.742Z")
-public class NotFoundException extends ApiException {
-	private int code;
-	public NotFoundException (int code, String msg) {
-		super(code, msg);
-		this.code = code;
-	}
-}
