+++ consonance-webservice/src/main/java/io/swagger/workflow/api/RunApi.java
@@ -1,7 +1,26 @@
-package io.swagger.api;
+/*
+ *     Consonance - workflow software for multiple clouds
+ *     Copyright (C) 2016 OICR
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package io.swagger.workflow.api;
 
 import io.swagger.annotations.ApiParam;
-import io.swagger.api.factories.RunApiServiceFactory;
+import io.swagger.workflow.api.factories.RunApiServiceFactory;
 
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
