+++ consonance-webservice/src/main/java/io/swagger/workflow/api/factories/RunApiServiceFactory.java
@@ -1,0 +1,34 @@
+/*
+ *     Consonance - workflow software for multiple clouds
+ *     Copyright (C) 2016 OICR
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package io.swagger.workflow.api.factories;
+
+import io.swagger.workflow.api.RunApiService;
+import io.swagger.workflow.api.impl.RunApiServiceImpl;
+
+@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2016-06-29T18:39:51.024Z")
+public class RunApiServiceFactory {
+
+   private final static RunApiService service = new RunApiServiceImpl();
+
+   public static RunApiService getRunApi()
+   {
+      return service;
+   }
+}
